Fixed bugs in the emulator

+++ src/constants.h
@@ -5,6 +5,6 @@
 #include <cstdint>
 
 constexpr size_t kMemorySize = 1024;
-using Word = int8_t;
+using Word = int16_t;
 
 #endif  // CONSTANTS_H_
